@@ -1,9 +1,9 @@
-import React, { useHistory } from 'react-router-dom';
+import React, { useNavigate } from 'react-router-dom';
 import './NavBar.component.scss';
 import logo from '../../assets/genesys-logo.png'
 
 export function NavBar() {
-    const history = useHistory();
+    const navigate = useNavigate();
 
     return (
             <div className="nav__wrapper">
@@ -13,13 +13,13 @@ export function NavBar() {
                 <div className="nav__link-section">
                     <ul className="nav">
                         <li className="nav__list-item">
-                            <button className="nav__link" onClick={() => history.push('')}>Home</button>
+                            <button className="nav__link" onClick={() => navigate('')}>Home</button>
                         </li>
                         <li className="nav__list-item">
-                            <button className="nav__link" onClick={() => history.push('user-search')}>User Search</button>
+                            <button className="nav__link" onClick={() => navigate('user-search')}>User Search</button>
                         </li>
                         <li className="nav__list-item">
-                            <button className="nav__link" onClick={() => history.push('/queues')}>Queues</button>
+                            <button className="nav__link" onClick={() => navigate('/queues')}>Queues</button>
                         </li>
                     </ul>
                 </div>
